added check for table to migration

+++ server/src/db/migrations/20220307225738_add-order-to-training-mod-items.ts
@@ -2,14 +2,25 @@ import { Knex } from "knex";
 
 
 export async function up(knex: Knex): Promise<void> {
-    return knex.schema.alterTable("Question", function (t) {
-        t.integer("order");
+
+    knex.schema.hasTable("Question").then(function (exists) {
+        if (!exists) return;
+
+        return knex.schema.alterTable("Question", function (t) {
+            t.integer("order");
+        });
     });
 }
 
+
 export async function down(knex: Knex): Promise<void> {
-    return knex.schema.alterTable("Question", function (t) {
-        t.dropColumn("order");
+
+    knex.schema.hasTable("Question").then(function (exists) {
+        if (!exists) return;
+
+        return knex.schema.alterTable("Question", function (t) {
+            t.dropColumn("order");
+        });
     });
 }
